% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_Wilmoth.R
\name{bifit}
\alias{bifit}
\title{Compute WLS using Tukey's biweight function}
\usage{
bifit(x, y, c = 6, tol = 1e-06, intercept = TRUE, tolerance = 1e-07,
  yname = NULL)
}
\arguments{
\item{x}{a matrix whose rows correspond to cases and whose columns correspond to variables.
x matrix should not include a column of ones}

\item{y}{the responses, possibly a matrix if you want to fit multiple left hand sides.
y can be matrix if there are multiple left-hand sides}

\item{c}{the tuning constant, typically around 6 to 9}

\item{tol}{the tolerance for convergence of the biweight fit}

\item{intercept}{logical. Whether or not an intercept term should be used.}

\item{tolerance}{the tolerance to be used in the matrix decomposition;}

\item{yname}{names to be used for the response variables.}
}
\description{
Compute WLS using Tukey's biweight function
}
\keyword{internal}
