% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_LinearLink.R
\name{check_LinearLink_input}
\alias{check_LinearLink_input}
\title{Check consistency in input arguments}
\usage{
check_LinearLink_input(input)
}
\arguments{
\item{input}{a list containing the input objects of the LinearLink function}
}
\description{
Check consistency in input arguments
}
\keyword{internal}
