% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_Wilmoth.R
\name{check_wilmoth_input}
\alias{check_wilmoth_input}
\title{Check wilmoth arguments}
\usage{
check_wilmoth_input(I)
}
\arguments{
\item{input}{a list containing the input objects of the wilmoth function}
}
\description{
Check wilmoth arguments
}
\keyword{internal}
