% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_LinearLink.R
\name{compute_lt_optim}
\alias{compute_lt_optim}
\title{Function to optimize a life table}
\usage{
compute_lt_optim(x, coefs, ex0)
}
\arguments{
\item{coefs}{coefficients}

\item{ex0}{life expectancy}

\item{ages}{ages}
}
\description{
Function to optimize a life table
}
\keyword{internal}
