% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_Wilmoth.R
\name{find.my.case}
\alias{find.my.case}
\title{Function that determintes the case/problem we have to solve
It also performes some checks}
\usage{
find.my.case(q0_5, k, e0, q15_45, q15_35, q0_1)
}
\description{
Function that determintes the case/problem we have to solve
It also performes some checks
}
\keyword{internal}
