% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_Wilmoth.R
\name{find_mx_optim}
\alias{find_mx_optim}
\title{Compute fitted mx}
\usage{
find_mx_optim(x, ex, x.f, coef, radix, k.int, verbose)
}
\arguments{
\item{x}{Numerical vector containing ages corresponding to the input data 
(in 'mx' or 'LT').}

\item{ex}{Life expectancy matrix}

\item{x.f}{Data frame containing the q0_5 values.}

\item{coef}{Estimated coefficients (ax, bx, cx) of the Log-Quadratic model.}

\item{radix}{Life table radix. Default: 10^5;}

\item{k.int}{a vector containing the end-points of the interval to be 
searched for k parameter.}

\item{verbose}{Logical. Choose whether to display a progress bar during the 
fitting process. Default: TRUE.}
}
\description{
Compute fitted mx
}
\keyword{internal}
