% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_LinearLink.R
\name{fitw_LSE}
\alias{fitw_LSE}
\title{Estimate bx using least square method and vx with SVD}
\usage{
fitw_LSE(log_ex_theta, log_mx, ...)
}
\description{
Estimate bx using least square method and vx with SVD
}
\keyword{internal}
