% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_A.r
\name{FUN.linearlink}
\alias{FUN.linearlink}
\title{Fit linear link model}
\usage{
FUN.linearlink(mx, mx_ages, mx_years, mx_country = NA, x_fit = 0,
  use.smooth = TRUE)
}
\description{
Fit linear link model
}
\keyword{internal}

