% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_lifetable.R
\name{FUN.lt_optim}
\alias{FUN.lt_optim}
\title{Function to optimize a life table}
\usage{
FUN.lt_optim(ages, coefs, ex0)
}
\arguments{
\item{ages}{ages}

\item{coefs}{coefficients}

\item{ex0}{life expectancy}
}
\value{
Results
}
\description{
Function to optimize a life table
}
\keyword{internal}

