% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_Kannisto.R
\name{Fun_ux}
\alias{Fun_ux}
\title{Select the mortality model}
\usage{
Fun_ux(model)
}
\description{
This function calls the mortality model that will be 
used in all the calculations. 
In this package only 1 model is implemented so far.
}
\keyword{internal}

