% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{HMD.test.data}
\alias{HMD.test.data}
\title{Death rates for female and male populations in US, France and Sweden}
\format{A list with 6 matrices with 101 rows and 50 columns:
\describe{
  \item{rows}{age interval}
  \item{colums}{years}
}}
\source{
Human Mortality Database, \url{http://www.mortality.org}.
}
\usage{
HMD.test.data
}
\description{
Dataset containing list containing 6 matrices with death rates for 
female and male populations in  USA, France and Sweden between 1965 and 2014. 
This data is provided for testing purposes only. It may not be updated.
Download the actual data free of charge from \url{http://www.mortality.org}.
}
\keyword{datasets}

