% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_imports.R
\docType{data}
\name{HMD3mx}
\alias{HMD3mx}
\title{Death Rates for Female and Male Populations in USA, France and Sweden}
\format{An object of class \code{list} of length 2.}
\source{
Human Mortality Database, \url{http://www.mortality.org}.
}
\usage{
HMD3mx
}
\description{
Dataset containing list containing 6 matrices with death rates for 
female and male populations in USA, France and Sweden between 1965 and 2014. 
This data is provided for testing purposes only. It may not be up to date anymore.
Download the actual data free of charge from \url{http://www.mortality.org}.
}
\keyword{datasets}
