% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MortalityEstimate-data.R
\docType{data}
\name{HMD4mx}
\alias{HMD4mx}
\title{Death Rates for Female Populations in England and Wales, France, Sweden and USA}
\format{An object of class \code{MortalityEstimateData} of length 4.}
\source{
Human Mortality Database, \url{http://www.mortality.org}.
}
\usage{
HMD4mx
}
\description{
Dataset containing containing 4 matrices with death rates for 
female populations in England and Wales, France, Sweden and USA 
between 1965 and 2014. This data is provided for testing purposes only. 
It may not be up to date anymore. Download the actual data free of charge 
from \url{http://www.mortality.org}.
}
\keyword{datasets}
