% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{HMD719}
\alias{HMD719}
\title{719 life tables from HMD (2012).}
\format{An object of class \code{data.frame} with 51768 rows and 12 columns.}
\usage{
HMD719
}
\description{
This data is provided for testing purposes only. 
This data is provided for testing purposes only. It may not be up to date anymore.
Download the actual data free of charge from \url{http://www.mortality.org}.
}
\keyword{datasets}
