% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{HMD_mx}
\alias{HMD_mx}
\title{Death rates for female populations in US, France and Sweden}
\format{A wide matrix with 101 rows and 49 columns:
\tabular{rl}{
Columns \tab -- Years (from 1965 to 2014)\cr
Rows \tab -- Ages (from 0 to 100)
}}
\source{
Human Mortality Database, \url{http://www.mortality.org}
}
\usage{
HMD_mx
}
\description{
A list containing 3 matrices with death rates for female populations in 
US, France, Sweden between 1965 and 2014.
}
\examples{
HMD_mx
}
\seealso{
\code{\link{HMD_LT}}
}
\keyword{datasets}

