% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_Kannisto.R
\name{Kannisto}
\alias{Kannisto}
\title{Fit Kannisto model for old age mortality}
\usage{
Kannisto(mx, x, parS = NULL)
}
\arguments{
\item{mx}{Matrix containing age-specific death rates (ages x years).
Can be also a vector or a data.frame with 1 column containing
rates in a single year}

\item{x}{Corresponding ages in the input matrix}

\item{parS}{Starting parameters used in optimization process}
}
\value{
A \code{Kannisto} object containing:

\item{fitted.values}{fitted values of the model}

\item{residuals}{residuals}

\item{coefficients}{Estimated coefficients}
}
\description{
This is the implementation of the Kannisto model for old age mortality. For more 
details see Thatcher et al. (1998).
}
\examples{

ages <- 80:100
mx  <- HMD3mx$female$SWE[paste(ages), ] # filter Sweden mx between age 80 and 100
fit_kan <- Kannisto(mx, x = ages) # fit Kannisto model
summary(fit_kan)

pred_kan  <- predict(fit_kan, 80:120) # extend mortality curve up to 120

}
\references{
Thatcher, A. R., Kannisto, V. and Vaupel, J. W. (1998) The Force 
of Mortality at Ages 80 to 120. Odense: Odense University Press. 104, 20 p. 
(Odense Monographs on Population Aging Vol. 5)
}
