% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_A.r
\name{life.table}
\alias{life.table}
\title{Function to create a life table with input variables: (age, Dx, Ex) 
or (age, mx) or (age, qx)}
\usage{
life.table(x, Dx = NULL, Ex = NULL, mx = NULL, qx = NULL, lx0 = 1e+05,
  ax0 = 0.1)
}
\arguments{
\item{x}{Ages}

\item{Dx}{Death counts}

\item{Ex}{Exposure}

\item{mx}{death rates}

\item{qx}{1 year death propabilities}

\item{lx0}{Radix}

\item{ax0}{0.1}
}
\value{
Results
}
\description{
Function to create a life table with input variables: (age, Dx, Ex) 
or (age, mx) or (age, qx)
}
\examples{
library(LinearLink)

mx_vector <- HMD_mx$USA[,1]
life.table(x = 0:100, mx = mx_vector)$lt

}

