% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_LinearLink.R
\name{LinearLinkLT}
\alias{LinearLinkLT}
\title{Estimate LinearLink Life Table}
\usage{
LinearLinkLT(object, ex, use.vx.rotation = FALSE, ...)
}
\arguments{
\item{object}{An object of class \code{LinearLink}.}

\item{ex}{Value of life expectancy for which we want to estimate 
the mortality curve. Type: numerical scalar.}

\item{use.vx.rotation}{Logical argument. If \code{TRUE} the adjustment method
inspired from \insertCite{li2013;textual}{MortalityEstimate} article is 
applied to the vx coefficients before estimated the life table. 
If \code{FALSE} the fitted vx coefficients are used in the estimations of 
the life table.}

\item{...}{Additional arguments affecting the predictions produced.}
}
\value{
Predicted values of the Linear-Link model.
}
\description{
Construct a life table based on the Linear-Link estimates and a given value 
of life expectancy at age theta 
(\insertCite{pascariu2020;textual}{MortalityEstimate}).
}
\examples{
# See examples in the ?LinearLink help page
}
\references{
\insertAllCited{}
}
\seealso{
\code{\link{LinearLink}}
}
