% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_LinearLink.R
\name{LSEfit}
\alias{LSEfit}
\title{Find the Least Squares Fit}
\usage{
LSEfit(x, y, intercept = F, tol.lsfit = 1e-07)
}
\arguments{
\item{x}{A matrix whose rows correspond to cases and whose columns correspond to variables.
x matrix should not include a column of ones}

\item{y}{The responses, possibly a matrix if you want to fit multiple left hand sides.
y can be matrix if there are multiple left-hand sides}

\item{intercept}{logical. Whether or not an intercept term should be used in 
least squares estimation. See \code{\link{lsfit}}.}

\item{tol.lsfit}{Tolerance to be used in the matrix decomposition in 
least squares estimation. See \code{\link{lsfit}}.}
}
\description{
Find the Least Squares Fit
}
\keyword{internal}
