% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_Wilmoth.R
\name{lthat.logquad}
\alias{lthat.logquad}
\title{Estimated life table using the log-quadratic model}
\usage{
lthat.logquad(coefs, x, sex, q0_5, k)
}
\description{
Estimated life table using the log-quadratic model
}
\keyword{internal}
