% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_imports.R
\docType{package}
\name{MortalityEstimate}
\alias{MortalityEstimate-package}
\title{Import Packages and Functions}
\description{
Indirect methods for estimating age-specific death rates.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/mpascariu/MortalityEstimate}
  \item Report bugs at \url{https://github.com/mpascariu/MortalityEstimate/issues}
}

}
\author{
\strong{Maintainer}: Marius D. Pascariu \email{rpascariu@outlook.com}

Other contributors:
\itemize{
  \item Jose Manuel Aburto [contributor]
  \item Ugofilippo Basellini [contributor]
  \item Vladimir Canudas-Romo [contributor]
}

}
