% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_Utils.R
\name{mx_qx}
\alias{mx_qx}
\title{Convert 'mx' into 'qx' and vice versa.}
\usage{
mx_qx(x, ux, out = "qx")
}
\arguments{
\item{ux}{a vector of mx or qx}
}
\description{
Function to convert mx into qx and back, using the constant force of 
mortality assumption (CFM).
}
\keyword{internal}
