% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_Wilmoth.R
\name{mxhat.logquad}
\alias{mxhat.logquad}
\title{Estimate mx using the log-quadratic model}
\usage{
mxhat.logquad(coefs, x, sex, q0_5, k = 0)
}
\description{
Estimate mx using the log-quadratic model
}
\keyword{internal}
