% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_LinearLink.R
\name{predict.LinearLink}
\alias{predict.LinearLink}
\title{Predict function for a LinearLink object}
\usage{
\method{predict}{LinearLink}(object, ex_target, use.vx.rotation = FALSE, ...)
}
\arguments{
\item{object}{An object of class 'LinearLink'}

\item{ex_target}{A value of life expectancy for which we want to estimate 
the mortality curve}

\item{use.vx.rotation}{Logical argument. If \code{TRUE} the adjustment method
described in Li et al. (2013) paper is applied to the vx coefficients before 
estimated the life table. If \code{FALSE} the fitted vx coefficients are used 
in the estimations of the life table.}

\item{...}{additional arguments affecting the predictions produced}
}
\value{
Predicted values of the Linear-Link model
}
\description{
Predict function for a LinearLink object
}
\references{
Li N., Lee R. and Gerland P. (2013). Extending the Lee-Carter 
Method to Model the Rotation of Age Patterns of Mortality Decline 
for Long-Term Projections. Demography 50:2037-2051. 
DOI: \url{http://dx.doi.org/10.1007/s13524-013-0232-2}
}
