% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_Wilmoth.R
\name{predict.wilmoth}
\alias{predict.wilmoth}
\title{Predict function for wilmoth object}
\usage{
\method{predict}{wilmoth}(object, Q5 = NULL, k = NULL, e0 = NULL,
  QQa = NULL, QQb = NULL, Q1 = NULL, tol = 1e-09, maxit = 200, ...)
}
\arguments{
\item{object}{An \code{\link{wilmoth}} object.}

\item{Q5}{5q0}

\item{k}{k}

\item{e0}{life expectancy at birth level}

\item{QQa}{45q15}

\item{QQb}{35q15}

\item{Q1}{1q0}

\item{tol}{tolerance level for convergence. The tolerance level, is relevant 
for case 12 and 13 (QQ and e0 are known)}

\item{maxit}{maximum number of iterations allowed. Default: 100.}

\item{...}{additional arguments affecting the predictions produced.}
}
\value{
life table(s) matching given inputs, and associated values of 
Q5, k, e0, QQa, QQb, and Q1
}
\description{
Predict function for wilmoth object
}
