% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MortalityEstimate-data.R
\name{print.MortalityEstimateData}
\alias{print.MortalityEstimateData}
\title{Print function for HMD4mx data}
\usage{
\method{print}{MortalityEstimateData}(x, ...)
}
\arguments{
\item{x}{A \code{MortalityEstimateData} object.}

\item{...}{Further arguments passed to or from other methods.}
}
\description{
Print function for HMD4mx data
}
\keyword{internal}
