% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_Wilmoth.R
\name{print.wilmoth}
\alias{print.wilmoth}
\title{Print function for wilmoth}
\usage{
\method{print}{wilmoth}(x, ...)
}
\arguments{
\item{x}{An object of class \code{wilmoth}.}

\item{...}{Further arguments passed to or from other methods.}
}
\description{
Print function for wilmoth
}
\keyword{internal}
