% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_Wilmoth.R
\name{summary.wilmoth}
\alias{summary.wilmoth}
\title{Summary for wilmoth}
\usage{
\method{summary}{wilmoth}(object, ...)
}
\arguments{
\item{object}{An object of class \code{wilmoth}.}

\item{...}{Further arguments passed to or from other methods.}
}
\description{
Summary for wilmoth
}
\keyword{internal}
