% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_Wilmoth.R
\name{wilmoth.control}
\alias{wilmoth.control}
\title{Auxiliary for Controlling \code{wilmoth} Fitting.}
\usage{
wilmoth.control(c = 6, intercept = TRUE, tol.biweight = 1e-06,
  tol.lsfit = 1e-07, lx0 = 10^5, k.int = c(-25, 20), nu = 1, nv = 1)
}
\arguments{
\item{c}{Tuning constant when compute weighted least squared using 
Tukey's biweight function. Typically, between 6 to 9.}

\item{intercept}{logical. Whether or not an intercept term should be used in 
least squares estimation. See \code{\link{lsfit}}.}

\item{tol.biweight}{Tolerance for convergence of the biweight fit}

\item{tol.lsfit}{Tolerance to be used in the matrix decomposition in 
least squares estimation. See \code{\link{lsfit}}.}

\item{lx0}{Life table radix. Default: 10^5.}

\item{k.int}{a vector containing the end-points of the interval to be 
searched for k parameter.}

\item{nu}{The number of left singular vectors to be computed in SVD method. 
This must between 0 and n = nrow(x). See \code{\link{svd}}.}

\item{nv}{the number of right singular vectors to be computed. 
This must be between 0 and p = ncol(x). See \code{\link{svd}}.}
}
\value{
List with control parameters.
}
\description{
Auxiliary for Controlling \code{wilmoth} Fitting.
}
\seealso{
\code{\link{wilmoth}}
}
