% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_Wilmoth.R
\name{wilmoth}
\alias{wilmoth}
\title{Fit Log-Quadratic model}
\usage{
wilmoth(x, mx = NULL, LT = NULL, sex)
}
\arguments{
\item{x}{Numerical vector containing ages covered in the life tables.
If abridge life tables are supplied in \code{data} then provide the lower 
bounds of the age intervals.}

\item{mx}{input data. A data.frame / matrix with death rates.}

\item{LT}{input data. A collection of life tables. If \code{mx} is provided 
\code{LT} is not necessary and viceversa.}

\item{sex}{Specify the sex of the population. Options: \code{"female", "male" or "total"}}
}
\value{
The output is of class \code{wilmoth} with the components:

\item{input}{ a list with input objects provided in the function}

\item{coefficients}{ estimated coefficients}

\item{fitted.values}{ fitted values}

\item{residuals}{ residuals}

\item{model_info}{ Model details (equation). The relationship between 
the death rate at age x, mx, and the probability of dying between birth and 
age 5.}
}
\description{
Estimating the log-quadratic model using the bi-weight procedure as 
described in the Appendix of Wilmoth et.al.(2012).
}
\examples{

# DATA
sex = "female"
HMD719f <- HMD719[HMD719$sex == sex, ]

# Fit model
x <- c(0,1, seq(5, 110, by = 5))
W <- wilmoth(x, LT = HMD719f, sex = sex)

}
\references{
John Wilmoth, Sarah Zureick, Vladimir Canudas-Romo, Mie Inoue & 
Cheryl Sawyer (2012): A flexible two-dimensional mortality model for use in 
indirect estimation, Population Studies: A Journal of Demography, 66:1, 1-28
\url{http://dx.doi.org/10.1080/00324728.2011.611411}
}
\seealso{
\code{\link{wilmothLT}}
}
