% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_Wilmoth.R
\name{bifit}
\alias{bifit}
\title{Compute WLS using Tukey's biweight function}
\usage{
bifit(x, y, c, intercept, tol.biweight, tol.lsfit)
}
\arguments{
\item{x}{A matrix whose rows correspond to cases and whose columns correspond to variables.
x matrix should not include a column of ones}

\item{y}{The responses, possibly a matrix if you want to fit multiple left hand sides.
y can be matrix if there are multiple left-hand sides}

\item{c}{Tuning constant when compute weighted least squared using 
Tukey's biweight function. Typically, between 6 to 9.}

\item{intercept}{logical. Whether or not an intercept term should be used in 
least squares estimation. See \code{\link{lsfit}}.}

\item{tol.biweight}{Tolerance for convergence of the biweight fit}

\item{tol.lsfit}{Tolerance to be used in the matrix decomposition in 
least squares estimation. See \code{\link{lsfit}}.}
}
\description{
Compute WLS using Tukey's biweight function
}
\keyword{internal}
