% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_Wilmoth.R
\name{check.wilmoth.input}
\alias{check.wilmoth.input}
\title{Check wilmoth argumetns}
\usage{
check.wilmoth.input(input)
}
\arguments{
\item{input}{a list containing the input objects of the wilmoth function}
}
\description{
Check wilmoth argumetns
}
\keyword{internal}
