% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_LinearLinkFit.R
\name{fitw_MLE}
\alias{fitw_MLE}
\title{# Estimate bx, vx and k with Poisson Likelihood Method}
\source{
Brouhns et al. 2002
}
\usage{
fitw_MLE(log_ex_theta, log_mx, ...)
}
\description{
The implemeted method of estimated the bx, vx and k coefficients of the 
LinearLink model is based on the approach described in Brouhns et al. 2002
for fittin the Lee-Carter model. 
Code writen by Jose Manuel Aburto with minor changes by Marius Pascariu
}
\keyword{internal}

