% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_Kannisto.R
\name{Kannisto}
\alias{Kannisto}
\title{Fit Kannisto model for old age mortality}
\usage{
Kannisto(mx, x, parS = NULL)
}
\arguments{
\item{mx}{Matrix containing age-specific death rates (ages x years).
Can be also a vector or a data.frame with 1 column containing
rates in a single year}

\item{x}{Corresponding ages in the input matrix}

\item{parS}{Starting parameters used in optimization process}
}
\value{
Results
}
\description{
This is a description
}
\examples{
library(LinearLink)

head(HMD.test.data$SWE) # check data for Sweden
ages <- 80:100
dta  <- HMD.test.data$SWE[paste(ages), ] # filter Sweden mx between age 80 and 100
fit_kan <- Kannisto(mx = dta, x = ages) # fit Kannisto model
summary(fit_kan)

pred_kan  <- predict(fit_kan, 80:120) # extend mortality curve up to 120

}

