% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_lifetable.R
\name{lifetable}
\alias{lifetable}
\title{Life table function}
\usage{
lifetable(x, Dx = NULL, Ex = NULL, mx = NULL, qx = NULL,
  sex = "total", lx0 = 1e+05)
}
\arguments{
\item{x}{Vector of ages}

\item{Dx}{Vector containing death counts. An element of the vector, Dx, 
represents the number of deaths during the year to persons aged x to x+1}

\item{Ex}{Vector containing the exposure in the period. 
Ex is the mid-year population aged x to x+1}

\item{mx}{Age-specific death rates}

\item{qx}{1-year probability of dying between age x and x+1}

\item{sex}{sex of the population considered here. This argument affects the 
first two values in the life table ax column. The values are computed based 
on Coale-Demeny method and are sligthly different for males than for females.}

\item{lx0}{Radix. Default: 100 000}
}
\value{
life tables
}
\description{
Function to create a life table with input variables: (age, Dx, Ex) 
or (age, mx) or (age, qx)
}
\examples{

# Example 1 --- Full life table --------------
F_mx <- HMD3mx$female$SWE
ages <- as.numeric(rownames(F_mx))
mx   <- F_mx[, paste(1970)]

LT = lifetable(x = ages, mx = mx, sex = "female")
LT

# Example 2 --- Abridge life table ------------
x  = c(0, 1, seq(5, 110, by = 5))
mx = c(0.05384, 0.00504, 0.00157, 0.00120, 0.00186, 0.00286, 0.00347, 0.00395, 
       0.00486, 0.00527, 0.00686, 0.00939, 0.01295, 0.01957, 0.03106, 0.04931, 
       0.08461, 0.12988, 0.18076, 0.23542, 0.30853, 0.39036, 0.47887, 0.55141)
lt = lifetable(x, mx = mx, sex = "female")
lt
}
