% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_LinearLink.R
\name{LinearLink}
\alias{LinearLink}
\title{Fit Linear-Link Model}
\usage{
LinearLink(x, mx, y, country = "...", theta = 0, use.smooth = TRUE,
  method = "MLE")
}
\arguments{
\item{x}{vector of ages corresponding to the mx matrix}

\item{mx}{Death rates matrix with age as row and time as column}

\item{y}{vector of years corresponding to the mx matrix}

\item{country}{Optional. The name of the country that the data 
corresponds to. The name is adopted in the output tables.}

\item{theta}{Age to be fitted}

\item{use.smooth}{Logical variable indicating whether the spline smoothing is 
applied or not to the estimated coefficients (bx and vx). The smoothing 
can be applied in order to avoid jumps in the mortality rates from one 
age to another. This using splines. One degree of freedom is allocated 
for every 5 year of age.}

\item{method}{Optimizing method. Least squared approach \code{LSE} or Poisson 
likelihood estimation \code{MLE}.}
}
\value{
A \code{LinearLink} object containing:

\item{input}{ a list with input objects provided in the function}

\item{coefficients}{ estimated coefficient}

\item{fitted.values}{ fitted values}

\item{residuals}{ estimated residuals}

\item{fitted.life.tables}{ life tables constructed using the fitted values}

\item{df_splines}{ degrees of freedom used in spline smoothing procedure}

\item{model_info}{ a description of the model}

\item{process_date}{ data and time stamp}
}
\description{
Fit Linear-Link Model
}
\examples{

# Select the 1965 - 1990 time interval and fit the Linear-Link model
ages  <- 0:100 # available ages in our datasets
years <- 1965:1990 # available years
sex   <- 'female'
SWEmx <- HMD3mx$female$SWE[paste(ages), paste(years)]

# Fit the Linear-Link using the least square approach (LSE).
M <- LinearLink(x  = ages,
                mx = SWEmx,
                y  = years,
                country = 'SWEDEN',
                theta   = 0,
                method  = 'LSE')
M
summary(M)
coef(M)
ls(M)


# Derive a mortality curve (life table) from a value of
# life expectancy at birth in 2014, say 84.05
e0 <- 84.05
LT1 <- LinearLinkLT(M, ex = e0)
LT2 <- LinearLinkLT(M, ex = e0, use.vx.rotation = TRUE) 

}
