% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_LinearLink.R
\name{LinearLink}
\alias{LinearLink}
\title{Fit the Linear-Link Model}
\usage{
LinearLink(
  x,
  mx,
  y,
  country = "...",
  theta = 0,
  use.smooth = TRUE,
  method = "LSE"
)
}
\arguments{
\item{x}{Numerical vector containing ages corresponding to the input 
data (mx).}

\item{mx}{Death rates matrix with age as row and time as column.}

\item{y}{Vector of years corresponding to the mx matrix.}

\item{country}{Optional. The name of the country that the data 
corresponds to. The name is adopted in the output tables.}

\item{theta}{Age to be fitted.}

\item{use.smooth}{Logical variable indicating whether the spline smoothing 
is applied or not to the estimated coefficients (bx and vx). The smoothing 
can be applied in order to avoid jumps in the mortality rates from one 
age to another. This using splines. One degree of freedom is allocated 
for every 5 year of age.}

\item{method}{Optimizing method. Least squared approach \code{LSE} or 
Poisson likelihood estimation \code{MLE} based on the approach described in 
\insertCite{brouhns2002;textual}{MortalityEstimate}. Default: \code{LSE}.}
}
\value{
A \code{LinearLink} object containing:
 \item{input}{List with input objects provided in the function}
 \item{coefficients}{Estimated coefficient}
 \item{fitted.values}{Fitted values}
 \item{residuals}{Estimated deviance residuals}
 \item{fitted.life.tables}{ Life tables constructed using the fitted values}
 \item{df_splines}{Degrees of freedom used in spline smoothing procedure}
 \item{model_info}{Description of the model}
 \item{process_date}{Data and time stamp}
}
\description{
Calibrate the the Linear-Link mortality model introduced in
\insertCite{pascariu2020;textual}{MortalityEstimate}.
}
\examples{
# Select the 1965 - 1990 time interval and fit the Linear-Link model
ages  <- 0:100 # available ages in our datasets
years <- 1965:1990 # available years
sex   <- 'female'
SWEmx <- HMD4mx$SWE[paste(ages), paste(years)]

# Fit the Linear-Link using the least square approach (LSE).
M <- LinearLink(x  = ages,
                mx = SWEmx,
                y  = years,
                country = 'SWEDEN',
                theta   = 0,
                method  = 'LSE')
M
summary(M)
coef(M)
ls(M)


# Derive a mortality curve (life table) from a value of
# life expectancy at birth in 2014, say 84.05
e0 <- 84.05
LT1 <- LinearLinkLT(M, ex = e0)
LT2 <- LinearLinkLT(M, ex = e0, use.vx.rotation = TRUE) 
}
\references{
\insertAllCited{}

Pascariu MD (2018). PhD Thesis: Modelling and Forecasting Mortality.
University of Southern Denmark, 53-70. URL:
\url{https://github.com/mpascariu/PhD-Thesis/blob/master/Thesis.pdf}.
}
