% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/A_Package.R, R/functions_A.r
\docType{package}
\name{LinearLink}
\alias{LinearLink}
\alias{LinearLink-package}
\title{LinearLink: A package for deriving age-specific death rates from life expectancy}
\usage{
LinearLink(mx, mx_ages, mx_years, mx_country = NA, x_fit = 0,
  use.smooth = TRUE)
}
\arguments{
\item{mx}{Death rates matrix with age as row and time as column}

\item{mx_ages}{vector of ages corresponding to the mx matrix}

\item{mx_years}{vector of years corresponding to the mx matrix}

\item{mx_country}{The name of the country that the data corresponds to.
The name will be adopted in the output tables. It is optional.}

\item{x_fit}{Age to be fitted}

\item{use.smooth}{Logical variable indicating wheter the spline smoothing is 
applyed or not to the estimated coefficients (bx and vx).}
}
\value{
Results
}
\description{
This package contains a method for deriving age-specific death rates from
life expectancy using the linear-link methodology

Fit Linear-Link model
}
\examples{
library(LinearLink)
library(dplyr)

# Select the 1965 - 1990 time interval and fit the Linear-Link model
ages    <- 0:100 # available ages in our datasets
years   <- 1965:1990 # available years
fit_SWE <- HMD_mx$SWE \%>\% select(mx.1965:mx.1990) \%>\%
     LinearLink(mx = ., ages, years, 'SWEDEN', x_fit = 0)
summary(fit_SWE)

# Derive a mortality curve (life table) from a value of
# life expectancy at birth in 2014, say 84.05
new_e0   <- 84.05
pred_SWE <- predict(fit_SWE, new_e0)
pred_SWE$lt
}

