% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_Wilmoth.R
\name{lt.from.mx}
\alias{lt.from.mx}
\title{Construct a life table from a vector of death rates}
\usage{
lt.from.mx(mx, x, sex, q0_5 = NULL, lx0 = 10^5)
}
\arguments{
\item{mx}{numeric vector of death dates}

\item{x}{numeric vector of ages.}

\item{sex}{sex. Possible values: \code{"female", "male"}, or \code{"total"}.}

\item{q0_5}{5q0}

\item{lx0}{radix. Default: 100 000.}
}
\description{
Construct a life table from a vector of death rates
}
\keyword{internal}
