% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_Wilmoth.R
\name{lthat.logquad}
\alias{lthat.logquad}
\title{Estimated life table using the log-quadratic model}
\usage{
lthat.logquad(coefs, x, q0_5, k, radix)
}
\arguments{
\item{coefs}{Estimated coefficients}

\item{x}{Numerical vector containing ages corresponding to the input data 
(in 'mx' or 'LT').}

\item{q0_5}{5q0. The probability that a new-born will die during the 
subsequent 5 years;}

\item{k}{k-parameter in the log-quadratic model;}

\item{radix}{Life table radix. Default: 10^5;}
}
\description{
Estimated life table using the log-quadratic model
}
\keyword{internal}
