% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MortalityEstimate-package.R
\docType{package}
\name{MortalityEstimate}
\alias{MortalityEstimate}
\alias{MortalityEstimate-package}
\title{Import Packages and Functions}
\description{
Indirect methods for estimating age-specific death rates.
}
\details{
This is a description. To be updated.
\insertNoCite{*}{MortalityEstimate}
}
\references{
\insertAllCited{}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/mpascariu/MortalityEstimate}
  \item Report bugs at \url{https://github.com/mpascariu/MortalityEstimate/issues}
}

}
\author{
\strong{Maintainer}: Marius D. Pascariu \email{rpascariu@outlook.com} (\href{https://orcid.org/0000-0002-2568-6489}{ORCID}) [copyright holder]

Authors:
\itemize{
  \item Jose Manuel Aburto (\href{https://orcid.org/0000-0002-2926-6879}{ORCID})
  \item Ugofilippo Basellini (\href{https://orcid.org/0000-0003-0292-1404}{ORCID})
  \item Vladimir Canudas-Romo (\href{https://orcid.org/0000-0001-6532-0089}{ORCID})
}

}
