% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MortalityEstimate-package.R
\docType{package}
\name{MortalityEstimate}
\title{Import Packages and Functions}
\description{
Indirect methods for estimating age-specific death rates.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/mpascariu/MortalityEstimate}
  \item Report bugs at \url{https://github.com/mpascariu/MortalityEstimate/issues}
}

}
\author{
\strong{Maintainer}: Marius D. Pascariu \email{rpascariu@outlook.com} (\href{https://orcid.org/0000-0002-2568-6489}{ORCID}) [copyright holder]

Authors:
\itemize{
  \item Jose Manuel Aburto
  \item Ugofilippo Basellini
  \item Vladimir Canudas-Romo
}

}
