% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_Wilmoth.R
\name{mxhat.logquad}
\alias{mxhat.logquad}
\title{Estimate mx using the log-quadratic model}
\usage{
mxhat.logquad(coefs, x, q0_5, k)
}
\arguments{
\item{x}{numeric vector of ages.}

\item{q0_5}{5q0}
}
\description{
Estimate mx using the log-quadratic model
}
\keyword{internal}
