% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_Generics.R
\name{predict.LinearLink}
\alias{predict.LinearLink}
\title{Predict function for a LinearLink object}
\source{
Li et al. (2013)
}
\usage{
\method{predict}{LinearLink}(object, ex_target, use.vx.rotation = FALSE, ...)
}
\arguments{
\item{object}{An object of class 'LinearLink'}

\item{ex_target}{A value of life expectancy for which we want to estimate 
the mortality curve}

\item{use.vx.rotation}{Logical argument. If \code{TRUE} the adjustment method
described in Li et al. (2013) paper is applied to the vx coefficients before 
estimated the life table. If \code{FALSE} the fitted vx coefficients are used 
in the estimations of the life table.}

\item{...}{Other arguments}
}
\description{
Predict function for a LinearLink object
}
