% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_Wilmoth.R
\name{predict.wilmoth}
\alias{predict.wilmoth}
\title{Predict function for wilmoth object}
\usage{
\method{predict}{wilmoth}(object, q0_5 = NULL, k = NULL, e0 = NULL,
  q15_45 = NULL, q15_35 = NULL, q0_1 = NULL, tol = 1e-09, maxit = 200,
  ...)
}
\arguments{
\item{object}{An \code{\link{wilmoth}} object.}

\item{q0_5}{5q0. The probability that a newborn will die during the subsequent 5 years}

\item{k}{k-parameter in the log-quadratic model}

\item{e0}{life expectancy at birth}

\item{q15_45}{45q15. The probability that a life aged 15 will die during the subsequent 45 years}

\item{q15_35}{35q15. The probability that a life aged 15 will die during the subsequent 35 years}

\item{q0_1}{1q0. The probability that a life aged 0 will die during the following year}

\item{tol}{tolerance level for convergence. The tolerance level, is relevant 
for case 12 and 13 (QQ and e0 are known)}

\item{maxit}{maximum number of iterations allowed. Default: 100.}

\item{...}{additional arguments affecting the predictions produced.}
}
\value{
a life table matching given inputs, and associated values of 
\code{5q0, k, e0, 45q15, 35q15,} and \code{1q0}.
}
\description{
Construct a life table with various choices of 2 input parameters 
\code{5q0, k, e0, 45q15, 35q15, 1q0} and a \code{wilmoth} object.
}
\examples{
# See examples in wilmoth()
}
\seealso{
\code{\link{wilmoth}}
}
