% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_LinearLink.R
\name{rotate_vx}
\alias{rotate_vx}
\title{Compute Rotated vx Coefficients}
\source{
Li et al. (2013)
}
\usage{
rotate_vx(object, ex_target, e0_u = 102, e0_threshold = 80,
  e0_conv = 130, p_ = 0.5)
}
\arguments{
\item{object}{An object of class 'LinearLink'}

\item{ex_target}{A value of life expectancy for which we want to derive the rotated vx}

\item{e0_u}{Ultimate value of life expectancy. At this point the rotation 
process reaches its maximum efficiency. Here. e0_u = 80 is taken as the default.}

\item{e0_threshold}{Level of life expectancy where the rotation should begin.
If rotated_vx is computed for ex_target <= e0_threshold then no difference
will be observed.}

\item{e0_conv}{Set limit for convergence. For life expectancy a birth the default value is 130.}

\item{p_}{The power to the smooth-weight function, p_, takes values 
between 0 and 1, which makes the rotation faster at starting times and 
slower at ending times. Here, p = .5 is taken as the default}
}
\description{
This functions computes the rotation of the vx coefficients using the method
presented in Li et al. (2013) paper.
}
\keyword{internal}
