% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_Wilmoth.R
\name{wilmoth}
\alias{wilmoth}
\title{Fit log-quadratic model using bi-weight method}
\usage{
wilmoth(data, sex, ages = c(0, 1, seq(5, 110, by = 5)))
}
\arguments{
\item{data}{input data. A collection of life tables.}

\item{sex}{Specify the sex of the population. Options: \code{"female", "male" or "total"}}

\item{ages}{Numerical vector containing ages covered in the life tables.
If abridge life tables are supplied in \code{data} then provide the lower 
bounds of the age intervals. Default: \code{c(0, 1, seq(5, 110, by = 5))} 
(corresponding to the HMD format).}
}
\value{
An \code{wilmoth} object
}
\description{
Fit log-quadratic model using bi-weight method
}
